+++ ts-client/src/dlmm/index.ts
@@ -25,7 +25,9 @@ import {
 import Decimal from "decimal.js";
 import {
   BASIS_POINT_MAX,
+  BIN_ARRAY_BITMAP_FEE_BN,
   BIN_ARRAY_FEE,
+  BIN_ARRAY_FEE_BN,
   LBCLMM_PROGRAM_IDS as DLMM_PROGRAM_IDS,
   FEE_PRECISION,
   MAX_ACTIVE_BIN_SLIPPAGE,
@@ -37,8 +39,10 @@ import {
   MAX_EXTRA_BIN_ARRAYS,
   MAX_FEE_RATE,
   POSITION_FEE,
+  POSITION_FEE_BN,
   PRECISION,
   SCALE_OFFSET,
+  TOKEN_ACCOUNT_FEE_BN,
   U64_MAX,
 } from "./constants";
 import { DlmmSdkError } from "./error";
@@ -92,6 +96,7 @@ import {
   compressBinAmount,
   computeBaseFactorFromFeeBps,
   distributeAmountToCompressedBinsByRatio,
+  findOptimumDecompressMultiplier,
   generateAmountForBinRange,
   getPositionCount,
   mulShr,
@@ -101,7 +106,8 @@ import {
   PositionV2Wrapper,
   getBinArrayAccountMetasCoverage,
   getBinArrayIndexesCoverage,
-  getPositionLowerUpperBinIdWithLiquidity,
+  isPositionNoFee,
+  isPositionNoReward,
   wrapPosition,
 } from "./helpers/positions";
 import {
@@ -159,6 +165,8 @@ import {
   CompressedBinDepositAmounts,
   BinLiquidityDistribution,
   LiquidityParameter,
+  SeedLiquidityCostBreakdown,
+  SeedLiquiditySingleBinResponse,
 } from "./types";
 import { bs58 } from "@coral-xyz/anchor/dist/cjs/utils/bytes";
 import { u64 } from "@coral-xyz/borsh";
@@ -2274,10 +2282,6 @@ export class DLMM {
 
     const [activeBin, positionsV2] = promiseResults;
 
-    const positions = [
-      ...positionsV2.map((p) => new PositionV2Wrapper(p.publicKey, p.account)),
-    ];
-
     if (!activeBin) {
       throw new Error("Error fetching active bin");
     }
@@ -2288,6 +2292,10 @@ export class DLMM {
         userPositions: [],
       };
     }
+
+    const positions = [
+      ...positionsV2.map((p) => new PositionV2Wrapper(p.publicKey, p.account)),
+    ];
 
     if (!positions) {
       throw new Error("Error fetching positions");
@@ -4170,7 +4178,8 @@ export class DLMM {
     const priceImpact = new Decimal(totalOutAmount.toString())
       .sub(new Decimal(outAmountWithoutSlippage.toString()))
       .div(new Decimal(outAmountWithoutSlippage.toString()))
-      .mul(new Decimal(100));
+      .mul(new Decimal(100))
+      .abs();
 
     const endPrice = getPriceOfBinByBinId(
       lastFilledActiveBinId.toNumber(),
@@ -4618,6 +4627,10 @@ export class DLMM {
     owner: PublicKey;
     position: LbPosition;
   }): Promise<Transaction> {
+    if (isPositionNoReward(position.positionData)) {
+      throw new Error("No LM reward to claim");
+    }
+
     const claimTransactions = await this.createClaimBuildMethod({
       owner,
       position,
@@ -4657,6 +4670,12 @@ export class DLMM {
     owner: PublicKey;
     positions: LbPosition[];
   }): Promise<Transaction[]> {
+    if (
+      positions.every((position) => isPositionNoReward(position.positionData))
+    ) {
+      throw new Error("No LM reward to claim");
+    }
+
     const claimAllTxs = (
       await Promise.all(
         positions
@@ -4757,7 +4776,11 @@ export class DLMM {
   }: {
     owner: PublicKey;
     position: LbPosition;
-  }): Promise<Transaction> {
+  }): Promise<Transaction | null> {
+    if (isPositionNoFee(position.positionData)) {
+      throw new Error("No fee to claim");
+    }
+
     const claimFeeTx = await this.createClaimSwapFeeMethod({ owner, position });
 
     const { blockhash, lastValidBlockHeight } =
@@ -4790,6 +4813,10 @@ export class DLMM {
     owner: PublicKey;
     positions: LbPosition[];
   }): Promise<Transaction[]> {
+    if (positions.every((position) => isPositionNoFee(position.positionData))) {
+      throw new Error("No fee to claim");
+    }
+
     const claimAllTxs = (
       await Promise.all(
         positions
@@ -4853,6 +4880,13 @@ export class DLMM {
     owner: PublicKey;
     position: LbPosition;
   }): Promise<Transaction[]> {
+    if (
+      isPositionNoFee(position.positionData) &&
+      isPositionNoReward(position.positionData)
+    ) {
+      throw new Error("No fee/reward to claim");
+    }
+
     const claimAllSwapFeeTxs = await this.createClaimSwapFeeMethod({
       owner,
       position,
@@ -4920,6 +4954,13 @@ export class DLMM {
     lockReleasePoint: BN,
     shouldSeedPositionOwner: boolean = false
   ): Promise<SeedLiquidityResponse> {
+    let tokenOwnerProveAssociatedTokenAccountLamports = new BN(0);
+    let totalPositionCount = new BN(0);
+    let totalPositionLamports = new BN(0);
+    let totalBinArraysCount = new BN(0);
+    let totalBinArraysLamports = new BN(0);
+    let binArrayBitmapLamports = new BN(0);
+
     const toLamportMultiplier = new Decimal(
       10 ** (this.tokenY.mint.decimals - this.tokenX.mint.decimals)
     );
@@ -4956,7 +4997,10 @@ export class DLMM {
       k
     );
 
-    const decompressMultiplier = new BN(10 ** this.tokenX.mint.decimals);
+    const decompressMultiplier = findOptimumDecompressMultiplier(
+      binDepositAmount,
+      new BN(this.tokenX.mint.decimals)
+    );
 
     let { compressedBinAmount, compressionLoss } = compressBinAmount(
       binDepositAmount,
@@ -4998,10 +5042,16 @@ export class DLMM {
       this.tokenX.owner
     );
 
+    const [binArrayBitmapExtension] = deriveBinArrayBitmapExtension(
+      this.pubkey,
+      this.program.programId
+    );
+
     const sendPositionOwnerTokenProveIxs = [];
     const initializeBinArraysAndPositionIxs = [];
     const addLiquidityIxs = [];
     const appendedInitBinArrayIx = new Set();
+    let appendedInitBinArrayBitmap = false;
 
     if (shouldSeedPositionOwner) {
       const positionOwnerTokenX =
@@ -5022,6 +5072,13 @@ export class DLMM {
       }
 
       if (requireTokenProve) {
+        if (!positionOwnerTokenX) {
+          tokenOwnerProveAssociatedTokenAccountLamports =
+            tokenOwnerProveAssociatedTokenAccountLamports.add(
+              TOKEN_ACCOUNT_FEE_BN
+            );
+        }
+
         const initPositionOwnerTokenX =
           createAssociatedTokenAccountIdempotentInstruction(
             payer,
@@ -5104,6 +5161,9 @@ export class DLMM {
         const index = binArrayIndexes[i];
 
         if (!account && !appendedInitBinArrayIx.has(pubkey)) {
+          totalBinArraysCount = totalBinArraysCount.add(new BN(1));
+          totalBinArraysLamports = totalBinArraysLamports.add(BIN_ARRAY_FEE_BN);
+
           instructions.push(
             await this.program.methods
               .initializeBinArray(index)
@@ -5119,6 +5179,9 @@ export class DLMM {
 
       const positionAccount = accounts.pop();
       if (!positionAccount) {
+        totalPositionCount = totalPositionCount.add(new BN(1));
+        totalPositionLamports = totalPositionLamports.add(POSITION_FEE_BN);
+
         instructions.push(
           await this.program.methods
             .initializePositionByOperator(
@@ -5143,7 +5206,7 @@ export class DLMM {
       }
 
       // Initialize bin arrays and initialize position account in 1 tx
-      if (instructions.length > 1) {
+      if (instructions.length > 0) {
         initializeBinArraysAndPositionIxs.push(instructions);
         instructions = [];
       }
@@ -5159,6 +5222,31 @@ export class DLMM {
           .gt(new BN(0));
 
       if (!positionDeposited) {
+        let overflowDefaultBinArrayBitmap = false;
+        for (const binArrayIndex of binArrayIndexes) {
+          if (isOverflowDefaultBinArrayBitmap(binArrayIndex)) {
+            if (!this.binArrayBitmapExtension && !appendedInitBinArrayBitmap) {
+              initializeBinArraysAndPositionIxs.push(
+                await this.program.methods
+                  .initializeBinArrayBitmapExtension()
+                  .accounts({
+                    binArrayBitmapExtension,
+                    funder: payer,
+                    lbPair: this.pubkey,
+                  })
+                  .instruction()
+              );
+
+              appendedInitBinArrayBitmap = true;
+              binArrayBitmapLamports = binArrayBitmapLamports.add(
+                BIN_ARRAY_BITMAP_FEE_BN
+              );
+            }
+
+            overflowDefaultBinArrayBitmap = true;
+          }
+        }
+
         const cappedUpperBinId = Math.min(
           upperBinId.toNumber(),
           maxBinId.toNumber() - 1
@@ -5188,8 +5276,8 @@ export class DLMM {
             .accounts({
               position: positionPda,
               lbPair: this.pubkey,
-              binArrayBitmapExtension: this.binArrayBitmapExtension
-                ? this.binArrayBitmapExtension.publicKey
+              binArrayBitmapExtension: overflowDefaultBinArrayBitmap
+                ? binArrayBitmapExtension
                 : this.program.programId,
               userToken: seederTokenX,
               reserve: this.lbPair.reserveX,
@@ -5234,8 +5322,8 @@ export class DLMM {
               .accounts({
                 position: positionPda,
                 lbPair: this.pubkey,
-                binArrayBitmapExtension: this.binArrayBitmapExtension
-                  ? this.binArrayBitmapExtension.publicKey
+                binArrayBitmapExtension: overflowDefaultBinArrayBitmap
+                  ? binArrayBitmapExtension
                   : this.program.programId,
                 userTokenX: seederTokenX,
                 userTokenY: seederTokenY,
@@ -5273,6 +5361,14 @@ export class DLMM {
       sendPositionOwnerTokenProveIxs,
       initializeBinArraysAndPositionIxs,
       addLiquidityIxs,
+      costBreakdown: {
+        tokenOwnerProveAssociatedTokenAccountLamports,
+        totalBinArraysCount,
+        totalBinArraysLamports,
+        totalPositionCount,
+        totalPositionLamports,
+        binArrayBitmapLamports,
+      },
     };
   }
 
@@ -5291,7 +5387,7 @@ export class DLMM {
    *    - `shouldSeedPositionOwner` (optional): Whether to send 1 lamport amount of token X to the position owner to prove ownership.
    *
    * The returned instructions need to be executed sequentially if it was separated into multiple transactions.
-   * @returns {Promise<TransactionInstruction[]>}
+   * @returns {Promise<SeedLiquiditySingleBinResponse>}
    */
   public async seedLiquiditySingleBin(
     payer: PublicKey,
@@ -5304,7 +5400,14 @@ export class DLMM {
     operator: PublicKey,
     lockReleasePoint: BN,
     shouldSeedPositionOwner: boolean = false
-  ): Promise<TransactionInstruction[]> {
+  ): Promise<SeedLiquiditySingleBinResponse> {
+    let tokenOwnerProveAssociatedTokenAccountLamports = new BN(0);
+    let totalPositionCount = new BN(0);
+    let totalPositionLamports = new BN(0);
+    let totalBinArraysCount = new BN(0);
+    let totalBinArraysLamports = new BN(0);
+    let binArrayBitmapLamports = new BN(0);
+
     const pricePerLamport = DLMM.getPricePerLamport(
       this.tokenX.mint.decimals,
       this.tokenY.mint.decimals,
@@ -5383,6 +5486,10 @@ export class DLMM {
             })
             .instruction()
         );
+
+        binArrayBitmapLamports = binArrayBitmapLamports.add(
+          BIN_ARRAY_BITMAP_FEE_BN
+        );
       }
     } else {
       binArrayBitmapExtension = this.program.programId;
@@ -5398,7 +5505,7 @@ export class DLMM {
       this.lbPair.tokenXMint,
       positionOwner,
       true,
-      this.tokenY.owner
+      this.tokenX.owner
     );
 
     if (shouldSeedPositionOwner) {
@@ -5439,10 +5546,10 @@ export class DLMM {
         const createPositionOwnerTokenXIx =
           createAssociatedTokenAccountIdempotentInstruction(
             payer,
+            positionOwnerTokenX,
             positionOwner,
             this.lbPair.tokenXMint,
-            this.tokenX.owner,
-            positionOwnerTokenX
+            this.tokenX.owner
           );
         preInstructions.push(createPositionOwnerTokenXIx);
 
@@ -5459,6 +5566,11 @@ export class DLMM {
         );
         transferIx.keys.push(...this.tokenX.transferHookAccountMetas);
         preInstructions.push(transferIx);
+
+        tokenOwnerProveAssociatedTokenAccountLamports =
+          tokenOwnerProveAssociatedTokenAccountLamports.add(
+            TOKEN_ACCOUNT_FEE_BN
+          );
       }
     }
 
@@ -5473,6 +5585,9 @@ export class DLMM {
           })
           .instruction()
       );
+
+      totalBinArraysCount = totalBinArraysCount.add(new BN(1));
+      totalBinArraysLamports = totalBinArraysLamports.add(BIN_ARRAY_FEE_BN);
     }
 
     if (!positionAccount) {
@@ -5496,6 +5611,9 @@ export class DLMM {
           })
           .instruction()
       );
+
+      totalPositionCount = totalPositionCount.add(new BN(1));
+      totalPositionLamports = totalPositionLamports.add(POSITION_FEE_BN);
     }
 
     const slices: RemainingAccountsInfoSlice[] = [
@@ -5541,8 +5659,8 @@ export class DLMM {
         tokenXMint: this.lbPair.tokenXMint,
         tokenYMint: this.lbPair.tokenYMint,
         sender: operator,
-        tokenXProgram: TOKEN_PROGRAM_ID,
-        tokenYProgram: TOKEN_PROGRAM_ID,
+        tokenXProgram: this.tokenX.owner,
+        tokenYProgram: this.tokenY.owner,
       })
       .remainingAccounts([
         ...transferHookAccountMetas,
@@ -5554,7 +5672,18 @@ export class DLMM {
       ])
       .instruction();
 
-    return [...preInstructions, depositLiquidityIx];
+    const instructions = [...preInstructions, depositLiquidityIx];
+    return {
+      instructions,
+      costBreakdown: {
+        tokenOwnerProveAssociatedTokenAccountLamports,
+        totalBinArraysCount,
+        totalBinArraysLamports,
+        totalPositionCount,
+        totalPositionLamports,
+        binArrayBitmapLamports,
+      },
+    };
   }
 
   /**
@@ -5825,12 +5954,18 @@ export class DLMM {
       this.lbPair,
       this.binArrayBitmapExtension?.account ?? null
     );
-    const accountsToFetch = [];
-    const [binArrayBitMapExtensionPubkey] = deriveBinArrayBitmapExtension(
-      this.pubkey,
-      this.program.programId
+    const marketPriceBinArrayIndex = binIdToBinArrayIndex(
+      new BN(marketPriceBinId)
     );
-    accountsToFetch.push(binArrayBitMapExtensionPubkey);
+    const accountsToFetch = [];
+    const binArrayBitMapExtensionPubkey = isOverflowDefaultBinArrayBitmap(
+      new BN(marketPriceBinArrayIndex)
+    )
+      ? deriveBinArrayBitmapExtension(this.pubkey, this.program.programId)[0]
+      : null;
+
+    binArrayBitMapExtensionPubkey &&
+      accountsToFetch.push(binArrayBitMapExtensionPubkey);
     const [fromBinArrayPubkey] = deriveBinArray(
       this.pubkey,
       fromBinArrayIndex,
@@ -5856,11 +5991,23 @@ export class DLMM {
         accountsToFetch
       );
 
+    const preInstructions: TransactionInstruction[] = [];
     let fromBinArray: PublicKey | null = null;
     let toBinArray: PublicKey | null = null;
     let binArrayBitmapExtension: PublicKey | null = null;
-    if (!!binArrayAccounts?.[0]) {
+    if (binArrayBitMapExtensionPubkey) {
       binArrayBitmapExtension = binArrayBitMapExtensionPubkey;
+      if (!binArrayAccounts?.[0]) {
+        const initializeBitmapExtensionIx = await this.program.methods
+          .initializeBinArrayBitmapExtension()
+          .accounts({
+            binArrayBitmapExtension: binArrayBitMapExtensionPubkey,
+            funder: owner,
+            lbPair: this.pubkey,
+          })
+          .instruction();
+        preInstructions.push(initializeBitmapExtensionIx);
+      }
     }
     if (!!binArrayAccounts?.[1]) {
       fromBinArray = fromBinArrayPubkey;
@@ -5880,6 +6027,7 @@ export class DLMM {
         fromBinArray,
         toBinArray,
       })
+      .preInstructions(preInstructions)
       .transaction();
 
     return new Transaction({
@@ -6245,7 +6393,7 @@ export class DLMM {
               feeAmountYPerTokenStored: ZERO,
               rewardPerTokenStored: [ZERO, ZERO],
               price: pricePerLamport,
-              version: binArray.version,
+              version: 2,
               pricePerToken: new Decimal(pricePerLamport)
                 .mul(new Decimal(10 ** (baseTokenDecimal - quoteTokenDecimal)))
                 .toString(),
@@ -6467,18 +6615,11 @@ export class DLMM {
     owner: PublicKey;
     position: LbPosition;
   }) {
-    // Avoid to attempt to load uninitialized bin array on the program
-    const maybeClaimableBinRange = getPositionLowerUpperBinIdWithLiquidity(
-      position.positionData
-    );
-
-    if (!maybeClaimableBinRange) return;
-
-    const { lowerBinId, upperBinId } = maybeClaimableBinRange;
+    const { lowerBinId, upperBinId } = position.positionData;
 
     const binArrayAccountsMeta = getBinArrayAccountMetasCoverage(
-      lowerBinId,
-      upperBinId,
+      new BN(lowerBinId),
+      new BN(upperBinId),
       this.pubkey,
       this.program.programId
     );
@@ -6501,7 +6642,7 @@ export class DLMM {
         this.getPotentialToken2022IxDataAndAccounts(ActionType.Reward, i);
 
       const claimTransaction = await this.program.methods
-        .claimReward2(new BN(i), lowerBinId.toNumber(), upperBinId.toNumber(), {
+        .claimReward2(new BN(i), lowerBinId, upperBinId, {
           slices,
         })
         .accounts({
@@ -6531,18 +6672,11 @@ export class DLMM {
     owner: PublicKey;
     position: LbPosition;
   }) {
-    // Avoid to attempt to load uninitialized bin array on the program
-    const maybeClaimableBinRange = getPositionLowerUpperBinIdWithLiquidity(
-      position.positionData
-    );
-
-    if (!maybeClaimableBinRange) return;
-
-    const { lowerBinId, upperBinId } = maybeClaimableBinRange;
+    const { lowerBinId, upperBinId } = position.positionData;
 
     const binArrayAccountsMeta = getBinArrayAccountMetasCoverage(
-      lowerBinId,
-      upperBinId,
+      new BN(lowerBinId),
+      new BN(upperBinId),
       this.pubkey,
       this.program.programId
     );
@@ -6591,7 +6725,7 @@ export class DLMM {
       this.getPotentialToken2022IxDataAndAccounts(ActionType.Liquidity);
 
     const claimFeeTx = await this.program.methods
-      .claimFee2(lowerBinId.toNumber(), upperBinId.toNumber(), {
+      .claimFee2(lowerBinId, upperBinId, {
         slices,
       })
       .accounts({